+++ CHANGELOG.md
@@ -40,6 +40,11 @@
       - Clean-up happens automatically, clients no longer need to call ```#clean_up``` (like previously mentioned).
       - Slave instances now buffer their logged issues and report them to the Master in batches.
       - ```#issues``` now returns the first variation of each issue to provide more info/context.
+  - ```Dispatcher```
+      - Added ```#workload_score``` returning the workload score of a Dispatcher as a ```Float```.
+      - Workload score calculation no longer uses CPU/RAM resource usage (since
+        that data is not available cross-platform) but instead the amount of running
+        instances and node weight.
 - Trainer -- Added a hard-limit for trainings per page to avoid time-consuming loops.
 - Spider
   - Updated to retry a few times when the server fails to respond.